# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BizQuestGender = typeof BizQuestGender[keyof typeof BizQuestGender];


export const BizQuestGender = {
  girl: 'girl',
  boy: 'boy',
  nonbinary: 'nonbinary',
  'prefer-not-to-say': 'prefer-not-to-say',
} as const;

export interface BizQuestRegisterRequest {
  /**
     * @minLength 3
     * @maxLength 18
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
  gender: BizQuestGender;
}

export interface BizQuestLoginRequest {
  /**
     * @minLength 3
     * @maxLength 18
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
}

export interface BizQuestAccount {
  id: string;
  username: string;
  gender: BizQuestGender;
}

export interface BizQuestAuthSession {
  token: string;
  user: BizQuestAccount;
}

export interface ApiError {
  message: string;
}