# lib/api-zod/src/generated/types/bizQuestAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BizQuestGender } from './bizQuestGender';

export interface BizQuestAccount {
  id: string;
  username: string;
  gender: BizQuestGender;
}