# lib/api-zod/src/generated/types/bizQuestAuthSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BizQuestAccount } from './bizQuestAccount';

export interface BizQuestAuthSession {
  token: string;
  user: BizQuestAccount;
}