# lib/api-zod/src/generated/types/bizQuestGender.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BizQuestGender = typeof BizQuestGender[keyof typeof BizQuestGender];


export const BizQuestGender = {
  girl: 'girl',
  boy: 'boy',
  nonbinary: 'nonbinary',
  'prefer-not-to-say': 'prefer-not-to-say',
} as const;