# lib/api-zod/src/generated/types/bizQuestLoginRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BizQuestLoginRequest {
  /**
     * @minLength 3
     * @maxLength 18
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
}