# lib/api-zod/src/generated/types/bizQuestRegisterRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BizQuestGender } from './bizQuestGender';

export interface BizQuestRegisterRequest {
  /**
     * @minLength 3
     * @maxLength 18
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 72
     */
  password: string;
  gender: BizQuestGender;
}